requirements.txt
fastapi
uvicorn
openai-whisper
transformers
torch
sentencepiece
spacy
pl-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pl_core_news_sm-3.5.0/pl_core_news_sm-3.5.0.tar.gz
